Update aarch64 tests
add define linux or android

+++ include/cpuinfo_aarch64.h
@@ -15,63 +15,66 @@
 #ifndef CPU_FEATURES_INCLUDE_CPUINFO_AARCH64_H_
 #define CPU_FEATURES_INCLUDE_CPUINFO_AARCH64_H_
 
+#include <stdint.h>
+
 #include "cpu_features_cache_info.h"
 #include "cpu_features_macros.h"
 
 CPU_FEATURES_START_CPP_NAMESPACE
 
 typedef struct {
-  int fp;          // Floating-point.
-  int asimd;       // Advanced SIMD.
-  int evtstrm;     // Generic timer generated events.
-  int aes;         // Hardware-accelerated Advanced Encryption Standard.
-  int pmull;       // Polynomial multiply long.
-  int sha1;        // Hardware-accelerated SHA1.
-  int sha2;        // Hardware-accelerated SHA2-256.
-  int crc32;       // Hardware-accelerated CRC-32.
-  int atomics;     // Armv8.1 atomic instructions.
-  int fphp;        // Half-precision floating point support.
-  int asimdhp;     // Advanced SIMD half-precision support.
-  int cpuid;       // Access to certain ID registers.
-  int asimdrdm;    // Rounding Double Multiply Accumulate/Subtract.
-  int jscvt;       // Support for JavaScript conversion.
-  int fcma;        // Floating point complex numbers.
-  int lrcpc;       // Support for weaker release consistency.
-  int dcpop;       // Data persistence writeback.
-  int sha3;        // Hardware-accelerated SHA3.
-  int sm3;         // Hardware-accelerated SM3.
-  int sm4;         // Hardware-accelerated SM4.
-  int asimddp;     // Dot product instruction.
-  int sha512;      // Hardware-accelerated SHA512.
-  int sve;         // Scalable Vector Extension.
-  int asimdfhm;    // Additional half-precision instructions.
-  int dit;         // Data independent timing.
-  int uscat;       // Unaligned atomics support.
-  int ilrcpc;      // Additional support for weaker release consistency.
-  int flagm;       // Flag manipulation instructions.
-  int ssbs;        // Speculative Store Bypass Safe PSTATE bit.
-  int sb;          // Speculation barrier.
-  int paca;        // Address authentication.
-  int pacg;        // Generic authentication.
-  int dcpodp;      // Data cache clean to point of persistence.
-  int sve2;        // Scalable Vector Extension (version 2).
-  int sveaes;      // SVE AES instructions.
-  int svepmull;    // SVE polynomial multiply long instructions.
-  int svebitperm;  // SVE bit permute instructions.
-  int svesha3;     // SVE SHA3 instructions.
-  int svesm4;      // SVE SM4 instructions.
-  int flagm2;      // Additional flag manipulation instructions.
-  int frint;       // Floating point to integer rounding.
-  int svei8mm;     // SVE Int8 matrix multiplication instructions.
-  int svef32mm;    // SVE FP32 matrix multiplication instruction.
-  int svef64mm;    // SVE FP64 matrix multiplication instructions.
-  int svebf16;     // SVE BFloat16 instructions.
-  int i8mm;        // Int8 matrix multiplication instructions.
-  int bf16;        // BFloat16 instructions.
-  int dgh;         // Data Gathering Hint instruction.
-  int rng;         // True random number generator support.
-  int bti;         // Branch target identification.
-  int mte;         // Memory tagging extension.
+  uint8_t fp;          // Floating-point.
+  uint8_t asimd;       // Advanced SIMD.
+  uint8_t evtstrm;     // Generic timer generated events.
+  uint8_t aes;         // Hardware-accelerated Advanced Encryption Standard.
+  uint8_t pmull;       // Polynomial multiply long.
+  uint8_t sha1;        // Hardware-accelerated SHA1.
+  uint8_t sha2;        // Hardware-accelerated SHA2-256.
+  uint8_t crc32;       // Hardware-accelerated CRC-32.
+  uint8_t atomics;     // Armv8.1 atomic instructions.
+  uint8_t fphp;        // Half-precision floating point support.
+  uint8_t asimdhp;     // Advanced SIMD half-precision support.
+  uint8_t cpuid;       // Access to certain ID registers.
+  uint8_t asimdrdm;    // Rounding Double Multiply Accumulate/Subtract.
+  uint8_t jscvt;       // Support for JavaScript conversion.
+  uint8_t fcma;        // Floating point complex numbers.
+  uint8_t lrcpc;       // Support for weaker release consistency.
+  uint8_t dcpop;       // Data persistence writeback.
+  uint8_t sha3;        // Hardware-accelerated SHA3.
+  uint8_t sm3;         // Hardware-accelerated SM3.
+  uint8_t sm4;         // Hardware-accelerated SM4.
+  uint8_t asimddp;     // Dot product instruction.
+  uint8_t sha512;      // Hardware-accelerated SHA512.
+  uint8_t sve;         // Scalable Vector Extension.
+  uint8_t asimdfhm;    // Additional half-precision instructions.
+  uint8_t dit;         // Data independent timing.
+  uint8_t uscat;       // Unaligned atomics support.
+  uint8_t ilrcpc;      // Additional support for weaker release consistency.
+  uint8_t flagm;       // Flag manipulation instructions.
+  uint8_t ssbs;        // Speculative Store Bypass Safe PSTATE bit.
+  uint8_t sb;          // Speculation barrier.
+  uint8_t paca;        // Address authentication.
+  uint8_t pacg;        // Generic authentication.
+  uint8_t dcpodp;      // Data cache clean to point of persistence.
+  uint8_t sve2;        // Scalable Vector Extension (version 2).
+  uint8_t sveaes;      // SVE AES instructions.
+  uint8_t svepmull;    // SVE polynomial multiply long instructions.
+  uint8_t svebitperm;  // SVE bit permute instructions.
+  uint8_t svesha3;     // SVE SHA3 instructions.
+  uint8_t svesm4;      // SVE SM4 instructions.
+  uint8_t flagm2;      // Additional flag manipulation instructions.
+  uint8_t frint;       // Floating point to integer rounding.
+  uint8_t svei8mm;     // SVE Int8 matrix multiplication instructions.
+  uint8_t svef32mm;    // SVE FP32 matrix multiplication instruction.
+  uint8_t svef64mm;    // SVE FP64 matrix multiplication instructions.
+  uint8_t svebf16;     // SVE BFloat16 instructions.
+  uint8_t i8mm;        // Int8 matrix multiplication instructions.
+  uint8_t bf16;        // BFloat16 instructions.
+  uint8_t dgh;         // Data Gathering Hint instruction.
+  uint8_t rng;         // True random number generator support.
+  uint8_t bti;         // Branch target identification.
+  uint8_t mte;         // Memory tagging extension.
+
   // Make sure to update Aarch64FeaturesEnum below if you add a field here.
 } Aarch64Features;
